chore(structure): reorganize components/pages into folders and fix imports
