+++ frontend/src/main.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import {BrowserRouter} from 'react-router-dom'
 import App from './App'
-import {AuthProvider} from './context/AuthContext.jsx'
+import {AuthProvider} from './context/AuthContext'
 import './styles/global.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
