+++ frontend/src/pages/CategoriesPage/CategoriesPage.module.css
