+++ frontend/src/pages/DetailPage/DetailPage.module.css
